No longer using get_locals and get_outers, removing JuliaVariables and MLStyle dependencies.
