+++ src/small_scripts.jl
@@ -45,40 +45,5 @@ function data_to_Dict(data::Union{Dict, DataFrame, DataFrameRow}, model::JuMP.Mo
     end
 end
 
-
-
-"""
-    get_locals(ex::Expr)
-
-Get variables in Expr that are locally defined.
-
-get_locals borrowed from here:
-https://discourse.julialang.org/t/determine-the-scope-of-symbols-in-an-expr/47993/2
-"""
-function get_locals(ex::Expr)
-    vars = (JuliaVariables.solve_from_local ∘ JuliaVariables.simplify_ex)(ex).args[1].bounds
-    map(x -> x.name, vars)
-end
-
-_get_outers(_) = Symbol[]
-_get_outers(x::JuliaVariables.Var) = x.is_global ? [x.name] : Symbol[]
-function _get_outers(ex::Expr)
-    MLStyle.@match ex begin
-        Expr(:(=), _, rhs) => _get_outers(rhs)
-        Expr(:tuple, _..., Expr(:(=), _, rhs)) => _get_outers(rhs)
-        Expr(_, args...) => mapreduce(_get_outers, vcat, args)
-    end
-end
-
-"""
-    get_outers(ex::Expr)
-
-Gets the Symbols in the expression that are not local.
-
-get_outers borrowed from here:
-https://discourse.julialang.org/t/determine-the-scope-of-symbols-in-an-expr/47993/2
-"""
-get_outers(ex) = (unique! ∘ _get_outers ∘ JuliaVariables.solve_from_local ∘ JuliaVariables.simplify_ex)(ex)
-
 """ Wrapper around Iterators.flatten. """
 flat(arr::Array) = collect(Iterators.flatten(arr))
